MdeModulePkg: Update the VariableLock shim to preserve interface
The VariableLock shim current fails if called twice because the underlying
VariablePolicy engine returns an error if a policy is set on an existing
variable.

This breaks existing code which expect it to silently pass if a variable
is locked multiple times (becuase it should "be locked").

Refactor the shim to confirm that the variable is indeed locked and then
change the error to EFI_SUCCESS.